tests/submission/PhilipTaylor/canvas/2d.imageData.object.properties.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.imageData.object.properties</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.imageData.html">imageData</a>.<a href="index.2d.imageData.object.html">object</a>.properties</h1>
<p class="desc">ImageData objects have the right properties</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.imageData.type">2d.imageData.type</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("ImageData objects have the right properties");
_addTest(function(canvas, ctx) {

var imgdata = ctx.getImageData(0, 0, 10, 10);
_assertEqual(typeof(imgdata.width), 'number', "typeof(imgdata.width)", "'number'");
_assertEqual(typeof(imgdata.height), 'number', "typeof(imgdata.height)", "'number'");
_assertEqual(typeof(imgdata.data), 'object', "typeof(imgdata.data)", "'object'");


});
</script>

<div id="log"></div>
</body></html>
